chatbot/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
requests>=2.26.0
mysql-connector-python>=8.0.0
google-genai>=1.0.0
